Change config ID from unsigned short to unsigned int so that a Vertex is exactly 16 bytes (multiple of 2)

+++ src/stdafx.h
@@ -6,7 +6,7 @@
 #define NOMINMAX
 #endif
 
-typedef unsigned short ConfigId;
+typedef unsigned int ConfigId;
 
 #ifdef USE_DOUBLE_PRECISION
